Extract LinkInteractor from LinkPaymentLauncher.
