+++ link/src/main/java/com/stripe/android/link/ui/inline/LinkInlineSignedIn.kt
@@ -1,5 +1,6 @@
 package com.stripe.android.link.ui.inline
 
+import androidx.annotation.RestrictTo
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Arrangement
@@ -22,20 +23,21 @@ import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
-import com.stripe.android.link.LinkPaymentLauncher
+import com.stripe.android.link.LinkInteractor
 import com.stripe.android.link.R
 import com.stripe.android.link.theme.linkShapes
 import com.stripe.android.uicore.StripeTheme
 import com.stripe.android.uicore.getBorderStroke
 import com.stripe.android.uicore.stripeColors
 
+@RestrictTo(RestrictTo.Scope.LIBRARY_GROUP)
 @Composable
 fun LinkInlineSignedIn(
-    linkPaymentLauncher: LinkPaymentLauncher,
+    linkInteractor: LinkInteractor,
     onLogout: () -> Unit,
     modifier: Modifier = Modifier
 ) {
-    linkPaymentLauncher.component?.let { component ->
+    linkInteractor.component?.let { component ->
         val viewModel: InlineSignupViewModel = viewModel(
             factory = InlineSignupViewModel.Factory(component.injector)
         )
