+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/PaymentOptionsViewModel.kt
@@ -14,6 +14,7 @@ import com.stripe.android.core.injection.Injectable
 import com.stripe.android.core.injection.Injector
 import com.stripe.android.core.injection.NonFallbackInjector
 import com.stripe.android.core.injection.injectWithFallback
+import com.stripe.android.link.LinkInteractor
 import com.stripe.android.model.PaymentIntent
 import com.stripe.android.payments.paymentlauncher.PaymentResult
 import com.stripe.android.paymentsheet.analytics.EventReporter
@@ -55,6 +56,7 @@ internal class PaymentOptionsViewModel @Inject constructor(
     lpmRepository: LpmRepository,
     savedStateHandle: SavedStateHandle,
     linkHandler: LinkHandler,
+    linkInteractor: LinkInteractor,
 ) : BaseSheetViewModel(
     application = application,
     config = args.state.config,
@@ -66,6 +68,7 @@ internal class PaymentOptionsViewModel @Inject constructor(
     lpmRepository = lpmRepository,
     savedStateHandle = savedStateHandle,
     linkHandler = linkHandler,
+    linkInteractor = linkInteractor,
     headerTextFactory = HeaderTextFactory(isCompleteFlow = false),
 ) {
