+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/BasePaymentSheetViewModelInjectionTest.kt
@@ -82,7 +82,7 @@ internal open class BasePaymentSheetViewModelInjectionTest {
         customerRepositoryPMs: List<PaymentMethod> = emptyList(),
         args: PaymentSheetContractV2.Args = PaymentSheetFixtures.ARGS_CUSTOMER_WITH_GOOGLEPAY
     ): PaymentSheetViewModel = runBlocking {
-        TestViewModelFactory.create { linkHandler, savedStateHandle ->
+        TestViewModelFactory.create { linkHandler, linkInteractor, savedStateHandle ->
             PaymentSheetViewModel(
                 ApplicationProvider.getApplicationContext(),
                 args,
@@ -116,6 +116,7 @@ internal open class BasePaymentSheetViewModelInjectionTest {
                 testDispatcher,
                 savedStateHandle = savedStateHandle,
                 linkHandler = linkHandler,
+                linkInteractor = linkInteractor,
                 intentConfirmationInterceptor = fakeIntentConfirmationInterceptor,
             )
         }
