+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/LinkHandlerTest.kt
@@ -6,6 +6,7 @@ import app.cash.turbine.test
 import app.cash.turbine.testIn
 import com.google.common.truth.Truth.assertThat
 import com.stripe.android.link.LinkActivityResult
+import com.stripe.android.link.LinkInteractor
 import com.stripe.android.link.LinkPaymentLauncher
 import com.stripe.android.link.model.AccountStatus
 import com.stripe.android.link.ui.inline.UserInput
@@ -196,7 +197,7 @@ class LinkHandlerTest {
 
         handler.processingState.test {
             ensureAllEventsConsumed() // Begin with no events.
-            whenever(linkLauncher.signInWithUserInput(any(), any()))
+            whenever(linkInteractor.signInWithUserInput(any(), any()))
                 .thenReturn(Result.success(true))
             testScope.launch {
                 handler.payWithLinkInline(userInput, cardSelection(), shouldCompleteLinkFlow)
@@ -232,7 +233,7 @@ class LinkHandlerTest {
 
         handler.processingState.test {
             ensureAllEventsConsumed() // Begin with no events.
-            whenever(linkLauncher.signInWithUserInput(any(), any()))
+            whenever(linkInteractor.signInWithUserInput(any(), any()))
                 .thenReturn(Result.success(true))
             testScope.launch {
                 handler.payWithLinkInline(userInput, cardSelection(), shouldCompleteLinkFlow)
@@ -265,7 +266,7 @@ class LinkHandlerTest {
         handler.processingState.test {
             accountStatusFlow.emit(AccountStatus.SignedOut)
             ensureAllEventsConsumed() // Begin with no events.
-            whenever(linkLauncher.signInWithUserInput(any(), any()))
+            whenever(linkInteractor.signInWithUserInput(any(), any()))
                 .thenReturn(Result.failure(IllegalStateException("Whoops")))
             testScope.launch {
                 handler.payWithLinkInline(userInput, cardSelection(), shouldCompleteLinkFlow)
@@ -335,9 +336,11 @@ private fun runLinkTest(
     testBlock: suspend LinkTestData.() -> Unit
 ): Unit = runTest {
     val linkLauncher = mock<LinkPaymentLauncher>()
+    val linkInteractor = mock<LinkInteractor>()
     val savedStateHandle = SavedStateHandle()
     val handler = LinkHandler(
         linkLauncher = linkLauncher,
+        linkInteractor = linkInteractor,
         savedStateHandle = savedStateHandle,
     )
     val processingStateTurbine = handler.processingState.testIn(backgroundScope)
@@ -352,13 +355,14 @@ private fun runLinkTest(
         shippingValues = null,
     )
 
-    whenever(linkLauncher.getAccountStatusFlow(eq(configuration))).thenReturn(accountStatusFlow)
+    whenever(linkInteractor.getAccountStatusFlow(eq(configuration))).thenReturn(accountStatusFlow)
 
     with(
         LinkTestDataImpl(
             testScope = this,
             handler = handler,
             linkLauncher = linkLauncher,
+            linkInteractor = linkInteractor,
             savedStateHandle = savedStateHandle,
             configuration = configuration,
             accountStatusFlow = accountStatusFlow,
@@ -391,6 +395,7 @@ private class LinkTestDataImpl(
     override val testScope: TestScope,
     override val handler: LinkHandler,
     override val linkLauncher: LinkPaymentLauncher,
+    override val linkInteractor: LinkInteractor,
     override val savedStateHandle: SavedStateHandle,
     override val configuration: LinkPaymentLauncher.Configuration,
     override val accountStatusFlow: MutableSharedFlow<AccountStatus>,
@@ -402,6 +407,7 @@ private interface LinkTestData {
     val testScope: TestScope
     val handler: LinkHandler
     val linkLauncher: LinkPaymentLauncher
+    val linkInteractor: LinkInteractor
     val savedStateHandle: SavedStateHandle
     val configuration: LinkPaymentLauncher.Configuration
     val accountStatusFlow: MutableSharedFlow<AccountStatus>
