+++ paymentsheet/src/test/java/com/stripe/android/paymentsheet/TestViewModelFactory.kt
@@ -1,22 +1,25 @@
 package com.stripe.android.paymentsheet
 
 import androidx.lifecycle.SavedStateHandle
-import com.stripe.android.link.LinkPaymentLauncher
+import com.stripe.android.link.LinkInteractor
 import com.stripe.android.paymentsheet.viewmodels.BaseSheetViewModel
 import org.mockito.kotlin.mock
 
 internal object TestViewModelFactory {
     fun <T : BaseSheetViewModel> create(
-        linkLauncher: LinkPaymentLauncher = mock(),
-        viewModelFactory: (linkHandler: LinkHandler, savedStateHandle: SavedStateHandle) -> T
+        linkInteractor: LinkInteractor = mock(),
+        viewModelFactory: (
+            linkHandler: LinkHandler,
+            linkInteractor: LinkInteractor,
+            savedStateHandle: SavedStateHandle,
+        ) -> T
     ): T {
-        lateinit var viewModel: T
         val savedStateHandle = SavedStateHandle()
         val linkHandler = LinkHandler(
-            linkLauncher = linkLauncher,
+            linkLauncher = mock(),
+            linkInteractor = linkInteractor,
             savedStateHandle = savedStateHandle,
         )
-        viewModel = viewModelFactory(linkHandler, savedStateHandle)
-        return viewModel
+        return viewModelFactory(linkHandler, linkInteractor, savedStateHandle)
     }
 }
